# docs/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "AnswersUpdate": {
        "additionalProperties": false,
        "properties": {
          "answers": {
            "additionalProperties": true,
            "title": "Answers",
            "type": "object"
          }
        },
        "required": [
          "answers"
        ],
        "title": "AnswersUpdate",
        "type": "object"
      },
      "AnswersValidation": {
        "properties": {
          "follow_ups": {
            "items": {
              "$ref": "#/components/schemas/FollowUpOut"
            },
            "title": "Follow Ups",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "valid": {
            "description": "True when the answers compile with no follow-up left.",
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "valid",
          "follow_ups"
        ],
        "title": "AnswersValidation",
        "type": "object"
      },
...
      "DiscoverySessionCreate": {
        "additionalProperties": false,
        "properties": {
          "organization": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Organization",
            "type": "string"
          },
          "template": {
            "enum": [
              "retail",
              "ministry",
              "bank"
            ],
            "title": "Template",
            "type": "string"
          }
        },
        "required": [
          "template",
          "organization"
        ],
        "title": "DiscoverySessionCreate",
        "type": "object"
      },
      "DiscoverySessionOut": {
        "properties": {
          "answers": {
            "additionalProperties": true,
            "title": "Answers",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "organization": {
            "title": "Organization",
            "type": "string"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "title": "Questions",
            "type": "array"
          },
          "resolved": {
            "title": "Resolved",
            "type": "boolean"
          },
          "template": {
            "title": "Template",
            "type": "string"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id"
          }
        },
        "required": [
          "id",
          "template",
          "organization",
          "questions",
          "answers",
          "resolved",
          "created_at"
        ],
        "title": "DiscoverySessionOut",
        "type": "object"
      },
      "FollowUpOut": {
        "properties": {
          "placeholder": {
            "title": "Placeholder",
            "type": "string"
          },
          "problem": {
            "title": "Problem",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "placeholder",
          "question",
          "problem"
        ],
        "title": "FollowUpOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QuestionOut": {
        "properties": {
          "placeholder": {
            "title": "Placeholder",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "placeholder",
          "question",
          "type"
        ],
        "title": "QuestionOut",
        "type": "object"
      },
...
      "ResolveRequest": {
        "additionalProperties": false,
        "properties": {
          "organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reuse an existing organization id instead of deriving one from the name.",
            "title": "Organization Id"
          }
        },
        "title": "ResolveRequest",
        "type": "object"
      },
      "ResolvedWorkflowOut": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "org_profile": {
            "additionalProperties": true,
            "title": "Org Profile",
            "type": "object"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "stage_count": {
            "title": "Stage Count",
            "type": "integer"
          },
          "template": {
            "title": "Template",
            "type": "string"
          },
          "workflow": {
            "additionalProperties": true,
            "title": "Workflow",
            "type": "object"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "organization_id",
          "workflow_id",
          "template",
          "name",
          "stage_count",
          "org_profile",
          "workflow"
        ],
        "title": "ResolvedWorkflowOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "medo433447@gmail.com",
      "name": "Mohamed Mahmoud"
    },
    "description": "\nCall the governed workflow engine from another system.\n\n**What this API will not do.** It will not approve anything on your behalf: a MEDIUM or\nHIGH stage moves only through `POST /runs/{run_id}/stages/{stage_key}/decision`, carrying a\nsigned human decision, and the authenticated principal is written into the hash-chained\nledger beside it. It will also not let a caller choose the model adapter \u2014 that is a\ndeployment setting (invariant **G7**).\n\n**Authentication.** Send `Authorization: Bearer <credential>`, where the credential is\neither a scoped API key (`atmpl_...`, minted with `atmpl keys create`) or a short-lived\naccess token from `POST /oauth/token`. Scopes are enforced per route and listed in each\nroute's errors.\n\n**Errors.** Every failure is `{\"error\": {\"code\", \"message\", \"details\"}}`. Every response\nechoes the `X-Request-Id` you sent, or one generated for you.\n\nAll demo organizations, people, and records reachable through this API are synthetic.\n",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Enterprise AI Automation Templates API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/discovery/sessions": {
      "post": {
        "description": "Returns the typed questionnaire and a blank, correctly shaped answers document.",
        "operationId": "create_session_discovery_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverySessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverySessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open a discovery session",
        "tags": [
          "discovery"
        ]
      }
    },
    "/discovery/sessions/{session_id}": {
      "get": {
        "operationId": "get_session_discovery_sessions__session_id__get",
        "parameters": [
          {
            "description": "Discovery session id.",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "description": "Discovery session id.",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverySessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a discovery session",
        "tags": [
          "discovery"
        ]
      }
    },
    "/discovery/sessions/{session_id}/answers": {
      "put": {
        "description": "Stores the answers and returns the outstanding follow-ups. An empty `follow_ups` list means the session can be resolved. Nothing compiles until then.",
        "operationId": "put_answers_discovery_sessions__session_id__answers_put",
        "parameters": [
          {
            "description": "Discovery session id.",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "description": "Discovery session id.",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswersUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswersValidation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit answers and see what is still missing",
        "tags": [
          "discovery"
        ]
      }
    },
    "/discovery/sessions/{session_id}/resolve": {
      "post": {
        "description": "Fails closed with `discovery_incomplete` and a numbered follow-up list while any typed answer is missing or invalid.",
        "operationId": "resolve_session_discovery_sessions__session_id__resolve_post",
        "parameters": [
          {
            "description": "Discovery session id.",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "description": "Discovery session id.",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedWorkflowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compile the session into a governed workflow",
        "tags": [
          "discovery"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This deployment",
      "url": "/api/v1"
    }
  ],
...
}